backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
sqlparse
psycopg2-binary
PyJWT
asgiref
pytz